Larastan and Duster
Had Duster fix the PostgresEngine.php so it would be happy.
Working through the new Larastan errors. No real code changes, but new type hinting to remove the errors.
Any ignored errors appear to be related to the fact that the code assumes that the Model is using the Searchable trait.